Added Widgets Override section
Hi,
In your documentation I did not I did not found how to do Widgets Override, How to rename it and how to assigned it. I did not able to rename so I did not write it. Only problem with rename.But my advice is to add documentation here about that.
Thanks
CEO
ThemeRox